Move BudgetList to model
